Complete VoltronPacket serialization refactor Packets can now have serializable types as properties TSOVoltronSerializer expanded for use outside of the TSOVoltronPacket.
