+++ nio2so.TSOTCP.City/TSO/Voltron/PDU/Datablob/TSOSimEventBroadcastPDU.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using nio2so.Formats.Streams;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -9,6 +10,7 @@ namespace nio2so.TSOTCP.City.TSO.Voltron.PDU.Datablob
     /// <summary>
     /// This is sent when the SimEvent kMSG is raised to the server
     /// </summary>
+    [TSOVoltronBroadcastDatablobPDU(TSO_PreAlpha_MasterConstantsTable.GZCLSID_cTSOSimEvent)]
     internal class TSOSimEventBroadcastPDU
     {
         [TSOVoltronBroadcastDatablobPDUField] public uint SimEvent_Arg1 { get; set; } = 0x00034D45;
@@ -17,8 +19,8 @@ namespace nio2so.TSOTCP.City.TSO.Voltron.PDU.Datablob
         [TSOVoltronBroadcastDatablobPDUField] public uint SimEvent_RefPack_Header_Arg2 { get; set; } = 0x7B6F2221;
         [TSOVoltronBroadcastDatablobPDUField] public uint SimEvent_RefPack_Header_Arg3 { get; set; } = 0x00000000;
 
-        //**Start RefPack**
+        //**Start TSOSerializableStream**
 
-        [TSOVoltronBroadcastDatablobPDUField] public byte RefPack_Byte { get; set; }
+        [TSOVoltronBroadcastDatablobPDUField] public TSOSerializableStream RefPackDataStream { get; set; }
     }
 }
