--- nio2so.TSOTCP.City/TSO/Voltron/Serialization/Struct/TSORefPackStream.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace nio2so.TSOTCP.City.TSO.Voltron.Serialization.Struct
-{
-    internal class TSORefPackStream
-    {
-    }
-}
