+++ nio2so.TSOTCP.City/nio2so.TSOTCP.City.csproj
@@ -18,4 +18,8 @@
     <ProjectReference Include="..\nio2so.Formats\nio2so.Formats.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="TSO\Voltron\Serialization\Struct\" />
+  </ItemGroup>
+
 </Project>
